+++ src/continuation/global_continuation_api.jl
@@ -18,7 +18,7 @@ abstract type GlobalContinuationAlgorithm end
 
 Find and continue attractors (or representations of attractors) and properties of their
 basins across a parameter curve `pcurve` according to algorithm `gca` and
-by sampling initial conditions using `icsampler`
+by sampling initial conditions using `icsampler`.
 
 Possible subtypes of a `GlobalContinuationAlgorithm` are:
 
@@ -53,33 +53,20 @@ global continuation. Use [`hilbert_pcurve`](@ref) to cover multiparameter spaces
 
 ## Description
 
-`global_continuation` is the central function of the framework for global stability analysis
-illustrated in [Datseris2023](@cite).
+`global_continuation` is the central function for performing global continuation
+as outlined in our article [Datseris2026](@cite).
 
-The global continuation algorithm typically references an [`BasinMap`](@ref)
+The global continuation algorithm typically references a [`BasinMap`](@ref)
 which is used to find the attractors and basins of a dynamical system. Additional
 arguments that control how to continue/track/match attractors across a parameter range
 are given when creating `gca`.
 
-The basin fractions and the attractors (or some representation of them) are continued
-across the parameter range `prange`, for the parameter of the system with index `pidx`
-(any index valid in `DynamicalSystems.set_parameter!` can be used). In contrast to
-traditional continuation (see online Tutorial for a comparison), global continuation
-can be performed over arbitrary user-defined curves in parameter space.
-The second call signature with `pcurve` allows for this possibility. In this case
-`pcurve` is a vector of iterables, where each iterable maps parameter indices
-to parameter values. These iterables can be dictionaries, named tuples, `Vector{Pair}`,
-anything that can be given in `set_parameters!`. We recommend to use dictionaries.
-Regardless, the sequence of the iterables defines a curve in parameter space.
-In fact, the version with `prange, pidx` simply defines
-`pcurve = [Dict(pidx => p) for p in prange]` and calls the second method.
+The basin properties and the attractors (or some representation of them) are continued
+across the parameter curve, whose elements are simply given to `DynamicalSystems.set_parameter!s`
+to update system parameters. This is fundamentally different to local continuation,
+see the online documentation or our article for details.
 """
-function global_continuation(alg::GlobalContinuationAlgorithm, prange::AbstractVector, pidx, sampler; kw...)
-    # everything is propagated to the curve setting
-    pcurve = [Dict(pidx => p) for p in prange]
-    return global_continuation(alg, pcurve, sampler; kw...)
-end
-
+function global_continuation end
 
 """
     continuation_series(continuation_info, fillval = NaN)